requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
duckduckgo-search>=6.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0  # For automatic retries on failure